requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.28
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.7
pydantic[email]
pydantic-settings>=2.2
passlib[bcrypt]>=1.7
PyJWT>=2.8
httpx>=0.27
pytest>=8.2
locust>=2.29
python-multipart